fix : interface코드이동
